+++ packages/create-hydrogen-app/CHANGELOG.md
@@ -9,6 +9,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 - Devcontainer support added [#164](https://github.com/Shopify/hydrogen/pull/164)
 - fix: add check for products.length in `Welcome.server.jsx`
+- BREAKING CHANGE: the previously default export from `@shopify/hydrogen/middleware` is now a named export `hydrogenMiddleware`.
 
 ## 0.6.4 - 2021-11-11